remove memory metrics
